move workflow exists function
