+++ cmd/workflow/deploy/registry_adapter_onchain.go
@@ -103,3 +103,19 @@ func (a *onchainRegistryAdapter) Upsert() error {
 	}
 	return nil
 }
+
+func (h *handler) workflowExists() error {
+	workflow, err := h.wrc.GetWorkflow(common.HexToAddress(h.settings.Workflow.UserWorkflowSettings.WorkflowOwnerAddress), h.inputs.WorkflowName, h.inputs.WorkflowTag)
+	if err != nil {
+		return err
+	}
+	if workflow.WorkflowId == [32]byte(common.Hex2Bytes(h.workflowArtifact.WorkflowID)) {
+		return fmt.Errorf("workflow with id %s already exists", h.workflowArtifact.WorkflowID)
+	}
+	if workflow.WorkflowName == h.inputs.WorkflowName {
+		status := workflow.Status
+		h.existingWorkflowStatus = &status
+		return fmt.Errorf("workflow with name %s already exists", h.inputs.WorkflowName)
+	}
+	return nil
+}
